fix(backend): restore the datasource list filter, and compile permissions once
The authorization filter this PR is half about was not actually in the branch.
Reverting the sqlc experiment with `git checkout -- backend/` took it with it,
while the frontend wording lives under app/ and survived -- so the panel told a
member "No shared connections you administrate" while the endpoint returned
every connection in the workspace. /simplify's efficiency pass caught it by
contradicting the premise I gave it.

Restored, and hoisted: Actor.CanManage recompiles the actor's permissions on
every call (MergeForRoles, then a filter, a convert and a core.Compile), which
is the wrong shape for a loop. Compiled once outside it, the per-row check is
the map lookup it should be. Same rule -- CanManage is exactly this expression.

Also from that pass: PatchStr, PatchStrDefault and PatchUUID had no callers left
once the generator stopped emitting them, PatchNullStr's comment pointed at one
of the deleted helpers, ds_cache had int(int64(x)) conversions left by the
migration, and a blanket rename in tables_test collapsed two assertions into a
duplicate, losing the check that a junction row emits no plain text patch.

+++ backend/internal/datasource/ds_cache.go
@@ -78,10 +78,10 @@ func GetOrLoadDatasource(ctx context.Context, id, workspaceID string) (*Resolved
 		Name:   row.Name,
 		DSN:    dsn,
 		Pool: engine.PoolConfig{
-			MaxOpenConns:    int(int64(row.MaxOpenConns)),
-			MaxIdleConns:    int(int64(row.MaxIdleConns)),
-			ConnMaxLifetime: time.Duration(int64(row.ConnMaxLifetime)) * time.Second,
-			ConnMaxIdleTime: time.Duration(int64(row.ConnMaxIdleTime)) * time.Second,
+			MaxOpenConns:    int(row.MaxOpenConns),
+			MaxIdleConns:    int(row.MaxIdleConns),
+			ConnMaxLifetime: time.Duration(row.ConnMaxLifetime) * time.Second,
+			ConnMaxIdleTime: time.Duration(row.ConnMaxIdleTime) * time.Second,
 		},
 	}
